make handle 64 bit

+++ src/XIVLauncher.Common/Dalamud/DalamudConsoleOutput.cs
@@ -4,7 +4,7 @@ namespace XIVLauncher.Common.Dalamud
     public sealed class DalamudConsoleOutput
     {
         public Int32 pid;
-        public Int32 handle;
+        public Int64 handle;
     }
 }
